+++ Arbiter.Net/Filters/ClientMessageFilter.cs
@@ -10,7 +10,7 @@ public delegate NetworkPacket? ClientMessageFilterHandler<TMessage>(ProxyConnect
 
 public class ClientMessageFilter<TMessage> : INetworkMessageFilter where TMessage : IClientMessage
 {
-    private readonly IClientMessageFactory _messageFactory = ClientMessageFactory.Default;
+    private readonly ClientMessageFactory _messageFactory = ClientMessageFactory.Default;
     private readonly ClientMessageFilterHandler<TMessage> _handler;
 
     public string? Name { get; init; }